move bluetooth enable
